+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/roomlist/room/RoomListElement.kt
@@ -4,18 +4,13 @@ import androidx.compose.animation.core.Spring
 import androidx.compose.animation.core.VisibilityThreshold
 import androidx.compose.animation.core.spring
 import androidx.compose.foundation.LocalIndication
-import androidx.compose.foundation.background
 import androidx.compose.foundation.gestures.detectTapGestures
 import androidx.compose.foundation.hoverable
 import androidx.compose.foundation.indication
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.interaction.PressInteraction
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyItemScope
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.LocalContentColor
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
@@ -23,14 +18,14 @@ import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.unit.IntOffset
-import androidx.compose.ui.unit.dp
 import de.connect2x.messenger.compose.view.DI
 import de.connect2x.messenger.compose.view.buttonPointerModifier
 import de.connect2x.messenger.compose.view.get
 import de.connect2x.messenger.compose.view.theme.components
+import de.connect2x.messenger.compose.view.theme.components.ThemedHorizontalDivider
+import de.connect2x.messenger.compose.view.theme.components.themedSurface
 import de.connect2x.trixnity.messenger.viewmodel.roomlist.RoomListElementViewModel
 import de.connect2x.trixnity.messenger.viewmodel.roomlist.RoomListViewModel
 import net.folivo.trixnity.core.model.RoomId
@@ -91,7 +86,6 @@ class RoomListElementContainerViewImpl : RoomListElementContainerView {
                 //         )
                 //      }
                 //  }
-                .indication(interactionSource, LocalIndication.current)
                 .hoverable(interactionSource, enabled = hoverable)
                 .pointerInput(roomName, roomId, isInvite) {
                     if (roomName == null || isInvite == null || isInvite == true) return@pointerInput
@@ -109,13 +103,18 @@ class RoomListElementContainerViewImpl : RoomListElementContainerView {
                         }
                     )
                 }
-                .background(if (roomId == selectedRoomId) { MaterialTheme.components.roomListSelection.color } else Color.Unspecified)
+                .then(if (roomId == selectedRoomId) Modifier.themedSurface(MaterialTheme.components.roomListSelection) else Modifier.themedSurface(MaterialTheme.components.roomListElement))
+                .indication(interactionSource, LocalIndication.current)
                 .buttonPointerModifier(enabled = isInvite == null || isInvite == false)
         ) {
-            CompositionLocalProvider(LocalContentColor provides if (roomId == selectedRoomId) MaterialTheme.components.roomListSelection.contentColor else LocalContentColor.current) {
+            CompositionLocalProvider(
+                LocalContentColor provides if (roomId == selectedRoomId) MaterialTheme.components.roomListSelection.contentColor else LocalContentColor.current
+            ) {
                 RoomListElement(roomListViewModel, roomListElementViewModel)
             }
         }
-        HorizontalDivider(Modifier.fillMaxWidth().width(1.dp).padding(horizontal = 10.dp))
+        ThemedHorizontalDivider(
+            style = MaterialTheme.components.roomListDivider
+        )
     }
 }